+++ packages/blocks/src/_common/transformers/index.ts
@@ -1,1 +1,3 @@
+export * as HtmlTransformer from './html.js';
+export * as MarkdownTransformer from './markdown.js';
 export * as ZipTransformer from './zip.js';
