+++ packages/blocks/src/_common/transformers/zip.ts
@@ -9,24 +9,7 @@ import type {
 import { getAssetName, Job } from '@blocksuite/store';
 import JSZip from 'jszip';
 
-import { replaceIdMiddleware } from './utils.js';
-
-export function createAssetsArchive(
-  assetsMap: Map<string, Blob>,
-  assetsIds: string[]
-) {
-  const zip = new JSZip();
-
-  const assets = zip.folder('assets');
-  assertExists(assetsMap);
-
-  assetsMap.forEach((blob, id) => {
-    if (!assetsIds.includes(id)) return;
-    assets?.file(getAssetName(assetsMap, id), blob);
-  });
-
-  return zip;
-}
+import { replaceIdMiddleware } from './middlewares.js';
 
 export async function exportPages(workspace: Workspace, pages: Page[]) {
   const zip = new JSZip();
