+++ packages/playground/apps/default/components/quick-edgeless-menu.ts
@@ -19,15 +19,15 @@ import '@shoelace-style/shoelace/dist/themes/dark.css';
 import {
   COLOR_VARIABLES,
   extractCssVariables,
-  HtmlAdapter,
+  HtmlTransformer,
+  MarkdownTransformer,
+  NOTE_WIDTH,
   ZipTransformer,
 } from '@blocksuite/blocks';
-import { NOTE_WIDTH } from '@blocksuite/blocks';
-import { MarkdownAdapter } from '@blocksuite/blocks';
 import type { ContentParser } from '@blocksuite/blocks/content-parser';
 import { ShadowlessElement } from '@blocksuite/lit';
 import type { AffineEditorContainer } from '@blocksuite/presets';
-import { Job, Utils, type Workspace } from '@blocksuite/store';
+import { Utils, type Workspace } from '@blocksuite/store';
 import type { SlTab, SlTabGroup } from '@shoelace-style/shoelace';
 import { setBasePath } from '@shoelace-style/shoelace/dist/utilities/base-path.js';
 import { css, html, nothing } from 'lit';
@@ -243,77 +243,11 @@ export class QuickEdgelessMenu extends ShadowlessElement {
   }
 
   private _exportHtml() {
-    const job = new Job({ workspace: this.workspace });
-    job.pageToSnapshot(window.page).then(snapshot => {
-      new HtmlAdapter()
-        .fromPageSnapshot({
-          snapshot,
-          assets: job.assetsManager,
-        })
-        .then(async result => {
-          let downloadBlob: Blob;
-          const element = document.createElement('a');
-          const contentBlob = new Blob([result.file], { type: 'plain/text' });
-          if (result.assetsIds.length > 0) {
-            const zip = ZipTransformer.createAssetsArchive(
-              job.assets,
-              result.assetsIds
-            );
-
-            zip.file('index.html', contentBlob);
-
-            downloadBlob = await zip.generateAsync({ type: 'blob' });
-            element.setAttribute('download', 'export.zip');
-          } else {
-            downloadBlob = contentBlob;
-            element.setAttribute('download', 'export.md');
-          }
-          const fileURL = URL.createObjectURL(downloadBlob);
-          element.setAttribute('href', fileURL);
-          element.style.display = 'none';
-          document.body.appendChild(element);
-          element.click();
-          document.body.removeChild(element);
-          URL.revokeObjectURL(fileURL);
-        });
-    });
+    HtmlTransformer.exportPage(this.page);
   }
 
   private _exportMarkDown() {
-    const job = new Job({ workspace: this.workspace });
-    job.pageToSnapshot(window.page).then(snapshot => {
-      new MarkdownAdapter()
-        .fromPageSnapshot({
-          snapshot,
-          assets: job.assetsManager,
-        })
-        .then(async result => {
-          let downloadBlob: Blob;
-          const element = document.createElement('a');
-          const contentBlob = new Blob([result.file], { type: 'plain/text' });
-          if (result.assetsIds.length > 0) {
-            const zip = ZipTransformer.createAssetsArchive(
-              job.assets,
-              result.assetsIds
-            );
-
-            zip.file('index.md', contentBlob);
-
-            downloadBlob = await zip.generateAsync({ type: 'blob' });
-            element.setAttribute('download', 'export.zip');
-          } else {
-            downloadBlob = contentBlob;
-            element.setAttribute('download', 'export.md');
-          }
-          const fileURL = URL.createObjectURL(downloadBlob);
-          element.setAttribute('href', fileURL);
-          element.style.display = 'none';
-          document.body.appendChild(element);
-          element.click();
-          document.body.removeChild(element);
-          URL.revokeObjectURL(fileURL);
-        });
-    });
+    MarkdownTransformer.exportPage(this.page);
   }
 
   private _exportPng() {
