+++ packages/playground/apps/starter/components/debug-menu.ts
@@ -21,8 +21,8 @@ import {
   createPage,
   extractCssVariables,
   FONT_FAMILY_VARIABLES,
-  HtmlAdapter,
-  MarkdownAdapter,
+  HtmlTransformer,
+  MarkdownTransformer,
   NOTE_WIDTH,
   SIZE_VARIABLES,
   type SurfaceBlockComponent,
@@ -40,7 +40,7 @@ import {
 import type { CopilotPanel } from '@blocksuite/presets';
 import { AffineEditorContainer } from '@blocksuite/presets';
 import type { BaseBlockModel } from '@blocksuite/store';
-import { Job, Utils, type Workspace } from '@blocksuite/store';
+import { Utils, type Workspace } from '@blocksuite/store';
 import type { SlDropdown } from '@shoelace-style/shoelace';
 import { setBasePath } from '@shoelace-style/shoelace/dist/utilities/base-path.js';
 import { css, html } from 'lit';
@@ -399,77 +399,11 @@ export class DebugMenu extends ShadowlessElement {
   }
 
   private _exportHtml() {
-    const job = new Job({ workspace: this.workspace });
-    job.pageToSnapshot(window.page).then(snapshot => {
-      new HtmlAdapter()
-        .fromPageSnapshot({
-          snapshot,
-          assets: job.assetsManager,
-        })
-        .then(async result => {
-          let downloadBlob: Blob;
-          const element = document.createElement('a');
-          const contentBlob = new Blob([result.file], { type: 'plain/text' });
-          if (result.assetsIds.length > 0) {
-            const zip = ZipTransformer.createAssetsArchive(
-              job.assets,
-              result.assetsIds
-            );
-
-            zip.file('index.html', contentBlob);
-
-            downloadBlob = await zip.generateAsync({ type: 'blob' });
-            element.setAttribute('download', 'export.zip');
-          } else {
-            downloadBlob = contentBlob;
-            element.setAttribute('download', 'export.md');
-          }
-          const fileURL = URL.createObjectURL(downloadBlob);
-          element.setAttribute('href', fileURL);
-          element.style.display = 'none';
-          document.body.appendChild(element);
-          element.click();
-          document.body.removeChild(element);
-          URL.revokeObjectURL(fileURL);
-        });
-    });
+    HtmlTransformer.exportPage(this.page);
   }
 
   private async _exportMarkDown() {
-    const job = new Job({ workspace: this.workspace });
-    job.pageToSnapshot(window.page).then(snapshot => {
-      new MarkdownAdapter()
-        .fromPageSnapshot({
-          snapshot,
-          assets: job.assetsManager,
-        })
-        .then(async result => {
-          let downloadBlob: Blob;
-          const element = document.createElement('a');
-          const contentBlob = new Blob([result.file], { type: 'plain/text' });
-          if (result.assetsIds.length > 0) {
-            const zip = ZipTransformer.createAssetsArchive(
-              job.assets,
-              result.assetsIds
-            );
-
-            zip.file('index.md', contentBlob);
-
-            downloadBlob = await zip.generateAsync({ type: 'blob' });
-            element.setAttribute('download', 'export.zip');
-          } else {
-            downloadBlob = contentBlob;
-            element.setAttribute('download', 'export.md');
-          }
-          const fileURL = URL.createObjectURL(downloadBlob);
-          element.setAttribute('href', fileURL);
-          element.style.display = 'none';
-          document.body.appendChild(element);
-          element.click();
-          document.body.removeChild(element);
-          URL.revokeObjectURL(fileURL);
-        });
-    });
+    MarkdownTransformer.exportPage(this.page);
   }
 
   private _exportPng() {
